Patch for Music Flashback
- Cleaned up files
- Fixed hand references
