+++ MonoSandbox/Components/HapticManager.cs
